+++ src/Delivery/Client.php
@@ -25,7 +25,7 @@ use Contentful\JsonHelper;
  */
 class Client extends BaseClient
 {
-    const VERSION = '2.1.0';
+    const VERSION = '2.2.0-dev';
 
     /**
      * @var ResourceBuilder