+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/contentful/contentful.php/compare/2.1.0...HEAD)
+## [2.2.0](https://github.com/contentful/contentful.php/tree/2.2.0) (2017-09-26)
 
 ### Changed
 * `DateHelper::formatForJson()` is now deprecated and will be removed in version 3. Use `Contentful\format_date_for_json()` instead.